security moved to nested classes inside config
