--- src/main/java/com/nekromant/telegram/model/Role.java
@@ -1,25 +1,0 @@
-package com.nekromant.telegram.model;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-import org.springframework.security.core.GrantedAuthority;
-
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-
-@NoArgsConstructor
-@AllArgsConstructor
-@Getter
-@Setter
-public class Role implements GrantedAuthority {
-
-    private String title;
-
-    @Override
-    public String getAuthority() {
-        return title;
-    }
-}
