+++ src/main/java/com/nekromant/telegram/model/UserInfo.java
@@ -2,12 +2,8 @@ package com.nekromant.telegram.model;
 
 import com.nekromant.telegram.contants.UserType;
 import lombok.*;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.userdetails.UserDetails;
 
 import javax.persistence.*;
-import java.util.Collection;
-import java.util.Set;
 
 @Entity
 @Data
@@ -15,7 +11,7 @@ import java.util.Set;
 @AllArgsConstructor
 @ToString
 @Builder
-public class UserInfo implements UserDetails {
+public class UserInfo {
     @Id
     private String userName;
 
@@ -27,49 +23,4 @@ public class UserInfo implements UserDetails {
 
     @Column
     private Boolean notifyAboutReports;
-
-    @Transient
-    private String password;
-
-    @Transient
-    private Set<Role> roles;
-
-    @Override
-    public Collection<? extends GrantedAuthority> getAuthorities() {
-        return roles;
-    }
-
-    @Override
-    public String getPassword() {
-        return password;
-    }
-
-    @Override
-    public String getUsername() {
-        return userName;
-    }
-
-    public String getUserName() {
-        return userName;
-    }
-
-    @Override
-    public boolean isAccountNonExpired() {
-        return true;
-    }
-
-    @Override
-    public boolean isAccountNonLocked() {
-        return true;
-    }
-
-    @Override
-    public boolean isCredentialsNonExpired() {
-        return true;
-    }
-
-    @Override
-    public boolean isEnabled() {
-        return true;
-    }
 }
