--- src/main/java/com/nekromant/telegram/service/UserDetailsService.java
@@ -1,8 +1,0 @@
-package com.nekromant.telegram.service;
-
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-
-public interface UserDetailsService {
-    UserDetails loadUserByUsername(String username) throws UsernameNotFoundException;
-}
